{{ cookiecutter.project_slug }}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.project_slug }}"
version = "{{ cookiecutter.version }}"
description = "{{ cookiecutter.description }}"
authors = ["{{ cookiecutter.author_name }} <{{ cookiecutter.author_email }}>"]
readme = "README.md"

[tool.poetry.extras]
mysql = ["aiomysql"]
sqlite = ["aiosqlite"]
pg = ["aiopg"]
all = ["aiomysql", "aiosqlite", "aiopg"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
sqlalchemy = "^2.0.35"
pydantic-settings = "^2.5.2"
orjson = "^3.10.7"
uvicorn = {version = "^0.31.1", extras = ["h11"]}
aiosqlite = {version = "^0.20.0", optional = true}
aiomysql = {version = "^0.2.0", optional = true}
aiopg = {version = "^1.4.0", optional = true}
python-jose = "^3.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.24.0"
pytest-randomly = "^3.15.0"
httpx = "^0.27.2"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = "-vvv --cov={{ cookiecutter.project_module }}  cov-report=term-missing"
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "session"

[tool.skjold]
sources = ["pyup"]
report_only = false
report_format = 'cli'
cache_dir = '.skjold_cache'
cache_expires = 86400
ignore_file = '.skjoldignore'
verbose = false
